Add silent install on bootstrap

+++ cu-compose/cu-docker-compose.sh
@@ -22,16 +22,18 @@
 function check-env {
   if [ -f .env ]; then
     source .env
-    echo "env File already exist would you like to use $CU_DOMAIN as domain name ? [y/n]"
-    if [ "$1" != "-y" ]; then
-        read PROD_ASW
-        if [ "$PROD_ASW" != "y" ] && [ "$PROD_ASW" != "n" ]; then
-            echo "Entrer y ou n!"
-            exit 1
-        elif [ "$PROD_ASW" = "n" ]; then
-            rm .env
-            generate-env
-        fi
+    if [ -n "$SILENT_INSTALL" ] || [ "$SILENT_INSTALL" = "no" ]; then
+      echo "env File already exist would you like to use $CU_DOMAIN as domain name ? [y/n]"
+      if [ "$1" != "-y" ]; then
+          read PROD_ASW
+          if [ "$PROD_ASW" != "y" ] && [ "$PROD_ASW" != "n" ]; then
+              echo "Entrer y ou n!"
+              exit 1
+          elif [ "$PROD_ASW" = "n" ]; then
+              rm .env
+              generate-env
+          fi
+      fi
     fi
   else
     generate-env
@@ -113,7 +115,7 @@ function generate-env {
   read MYSQL_DATABASE
   if [ -n "$MYSQL_DATABSE" ]; then
     echo "MYSQL_DATABASE=$MYSQL_DATABASE" >> .env
-  else 
+  else
     echo "MYSQL_DATABASE=cloudunit" >> .env
   fi
   #echo "Which git branch would you want to deploy : [ default to 'dev' ]"
